kinit-api/requirements.txt
alembic==1.7.5
anyio==3.5.0
asgiref==3.5.0
asyncmy==0.2.5
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.0.12
click==8.0.3
colorama==0.4.4
cryptography==37.0.4
ecdsa==0.18.0
fastapi==0.73.0
greenlet==1.1.2
gunicorn==20.1.0
h11==0.13.0
httptools==0.3.0
idna==3.3
importlib-metadata==4.10.1
importlib-resources==5.4.0
loguru==0.5.3
Mako==1.1.6
MarkupSafe==2.0.1
orjson==3.6.6
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.0
PyMySQL==1.0.2
python-dotenv==0.19.2
python-jose==3.3.0
PyYAML==6.0
requests==2.28.0
rsa==4.9
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.31
SQLAlchemy-Utils==0.38.3
starlette==0.17.1
typing-extensions==4.0.1
ua-parser==0.15.0
urllib3==1.26.9
user-agents==2.2.0
uvicorn==0.17.0.post1
watchgod==0.7
websockets==10.1
win32-setctime==1.1.0
zipp==3.7.0
